Refactor: Add AppLayout component and use it in every page that an auth user can access. Remove logic that controls if an user is authenticated from App.vue
