+++ src/components/AppLayout.vue
@@ -1,0 +1,10 @@
+<script setup>
+import Navbar from '../components/Navbar.vue';
+
+</script>
+<template>
+    <div>
+        <Navbar></Navbar>
+        <slot></slot>
+    </div>
+</template>
